Move more specific functionality to the play namespace.
The core namespace no longer has to read the files during the site map creation pass. We still do for the purposes of the example. All that is required is some 'info' function to be passed as an argument to compilation. This function should return a map, and can choose to override the :url key of the output page.